+++ version/changelog.md
@@ -7,7 +7,7 @@
 - add progressbar when zipping files
 - add progressbar when copying files
 - support cancel for file operations (they did not cancel in backend before)
--
+- build asar archives in build (less files in binary)
 
 ### v1.2.0 (27.01.2026)